pyproject.toml
[project]
name = "sat_data_acquisition"
version = "0.1.0"
description = "Lightweight satellite data retrieval from public STAC sources"
readme = "README.md"
authors = [
    { name = "Peter Kongstad" }
]
keywords = [
    "satellite",
    "imagery",
    "raster",
    "STAC",
    "Sentinel",
    "Landsat",
    "geospatial",
]
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "geopandas>=0.13.0",
    "rasterio>=1.3.0",
    "xarray>=2023.1.0",
    "odc-stac>=0.3.0",
    "pystac-client>=0.6.0",
    "planetary-computer>=0.4.9",
    "rioxarray>=0.14.0",
    "shapely>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyproj>=3.4.0",
    "utm>=0.7.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.28.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "pytest-cov>=4.0.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: GIS",
]

[project.urls]
Homepage = "https://github.com/yourusername/sat_data_acquisition"
Repository = "https://github.com/yourusername/sat_data_acquisition"
Documentation = "https://github.com/yourusername/sat_data_acquisition#readme"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = { find = {} }

[tool.setuptools.package-data]
sat_data_acquisition = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
filterwarnings = [
    "ignore::DeprecationWarning:pydantic._internal._config",
    "ignore::pydantic.warnings.PydanticDeprecatedSince20:planetary_computer.sas",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["sat_data_acquisition"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
known-first-party = ["sat_data_acquisition"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
